+++ designer/client/src/components/graph/NodeDescriptionPopover.tsx
@@ -12,34 +12,7 @@ import { getNodeData } from "./Graph";
 import { isStickyNoteElement } from "./GraphPartialsInTS/cellUtils";
 import { MarkdownStyled } from "./node-modal/MarkdownStyled";
 import { Events } from "./types";
-
-const watchForCover = (getElement: () => Element, callback: (covered: boolean) => void) => {
-    let lastState = false;
-
-    const observer = new MutationObserver(() => {
-        const element = getElement();
-        if (!element) return;
-
-        const rect = element.getBoundingClientRect();
-        const centerX = rect.left + rect.width / 2;
-        const centerY = rect.top + rect.height / 2;
-        const topElement = document.elementFromPoint(centerX, centerY);
-        const isCovered = topElement !== null && topElement !== element && !element.contains(topElement);
-
-        if (isCovered !== lastState) {
-            callback(isCovered);
-            lastState = isCovered;
-        }
-    });
-
-    observer.observe(document.body, {
-        attributes: true,
-        childList: true,
-        subtree: true,
-    });
-
-    return () => observer.disconnect();
-};
+import { watchForCover } from "./watchForCover";
 
 const useTimeout = <A extends Array<unknown>>(
     callback: (...args: A) => void,