+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/test/testcase/CompiledAssertions.scala
@@ -1,0 +1,8 @@
+package pl.touk.nussknacker.ui.process.test.testcase
+
+import pl.touk.nussknacker.engine.api.NodeId
+import pl.touk.nussknacker.engine.expression.parse.CompiledExpression
+
+final case class CompiledAssertions(assertions: Map[NodeId, List[CompiledAssertion]])
+
+final case class CompiledAssertion(expression: CompiledExpression)